app/requirements.txt
Flask
Gunicorn
flask
flask_cors
google.oauth2.service_account
gspread